Safari-611.1.10/LayoutTests/ietestcenter/Javascript/11.1.5_4-4-c-2-expected.txt

11.1.5_4-4-c-2 (TestCases/chapter11/11.1/11.1.5/11.1.5_4-4-c-2.js) Object literal - SyntaxError if a set accessor property definition is followed by a data property definition with the same name

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
